Modified and improved Models

+++ payroll_system/db/migrate/20230216105835_create_tax_deductions.rb
@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 class CreateTaxDeductions < ActiveRecord::Migration[7.0]
   def change
     create_table :tax_deductions do |t|
       t.string :deduct_type
-      t.float :amount_deduct
+      t.decimal :amount_deduct
       t.date :financial_year
 
       t.timestamps
