+++ payroll_system/db/migrate/20230216110251_create_payslips.rb
@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 class CreatePayslips < ActiveRecord::Migration[7.0]
   def change
     create_table :payslips do |t|
-      t.float :attendance_deductible
-      t.float :taxable_salary
-      t.float :tot_tax
-      t.float :payable_salary
+      t.decimal :attendance_deductible
+      t.decimal :taxable_salary
+      t.decimal :tot_tax
+      t.decimal :payable_salary
       t.date :slip_month
 
       t.timestamps
